musli
Module alloc

Allocation support for Müsli.

This crate contains two types of allocators:

The System allocator, which uses the system allocation facilities. Particularly std::alloc::System.

The Slice allocator, which can allocate buffers from a fixed-size slice.

The following types are also provided for convenience:

Vec which can be used as a vector of allocations.

String which can be used as a safe string container.

Examples

use musli::alloc::Vec;

musli::alloc::default!(|alloc| {
    let mut a = Vec::new_in(alloc);
    let mut b = Vec::new_in(alloc);

    b.write(b"He11o");
    a.write(b.as_slice());

    assert_eq!(a.as_slice(), b"He11o");
    assert_eq!(a.len(), 5);

    a.write(b" W0rld");

    assert_eq!(a.as_slice(), b"He11o W0rld");
    assert_eq!(a.len(), 11);

    let mut c = Vec::new_in(alloc);
    c.write(b"!");
    a.write(c.as_slice());

    assert_eq!(a.as_slice(), b"He11o W0rld!");
    assert_eq!(a.len(), 12);
});

Explicitly using a buffer on the stack with the Slice allocator:

use musli::alloc::{ArrayBuffer, Slice, Vec};

let mut buf = ArrayBuffer::new();
let alloc = Slice::new(&mut buf);

let mut a = Vec::new_in(&alloc);
let mut b = Vec::new_in(&alloc);

b.write(b"He11o");
a.write(b.as_slice());

assert_eq!(a.as_slice(), b"He11o");
assert_eq!(a.len(), 5);

a.write(b" W0rld");

assert_eq!(a.as_slice(), b"He11o W0rld");
assert_eq!(a.len(), 11);

let mut c = Vec::new_in(&alloc);
c.write(b"!");
a.write(c.as_slice());

assert_eq!(a.as_slice(), b"He11o W0rld!");
assert_eq!(a.len(), 12);

Macros

default

Call the given block $body with the default allocator.

Structs

ArrayBuffer

An array that can conveniently be used as a buffer, by default this is DEFAULT_ARRAY_BUFFER bytes large.

Disabled

An allocator which cannot allocate anything.

Slice

A no-std compatible slice-based allocator that can be used with the musli crate.

String

Wrapper around a buffer that is guaranteed to be a valid utf-8 string.

System

System buffer that can be used in combination with an Allocator.

Vec

A vector backed by an Allocator.

Constants

DEFAULT_ARRAY_BUFFER

The default stack buffer size for the default allocator provided through default!.

Statics

SYSTEM

The static system allocator instance.

Traits

Allocator

An allocator that can be used in combination with a context.

RawVec

A raw buffer allocated through an Allocator.