musli::alloc
Macro default

macro_rules! default {
    (|$alloc:ident| $body:block) => { ... };
}

Call the given block $body with the default allocator.

This is useful if you want to write application which are agnostic to whether the alloc feature is or isn’t enabled.

If the alloc feature is enabled, this is the System allocator.

If the alloc feature is disabled, this is the Slice allocator with DEFAULT_ARRAY_BUFFER bytes allocated on the stack.

Examples

use musli::alloc::Vec;

musli::alloc::default!(|alloc| {
    let mut a = Vec::new_in(alloc);
    let mut b = Vec::new_in(alloc);

    b.write(b"He11o");
    a.write(b.as_slice());

    assert_eq!(a.as_slice(), b"He11o");
    assert_eq!(a.len(), 5);

    a.write(b" W0rld");

    assert_eq!(a.as_slice(), b"He11o W0rld");
    assert_eq!(a.len(), 11);

    let mut c = Vec::new_in(alloc);
    c.write(b"!");
    a.write(c.as_slice());

    assert_eq!(a.as_slice(), b"He11o W0rld!");
    assert_eq!(a.len(), 12);
});