musli::compat

Struct Packed

#[repr(transparent)]
pub struct Packed<T>(pub T);

Treat T as if its packed.

This corresponds to the “Bytes” type in the [data model of Müsli]. It encodes any Encode / Decode type “on after another” and is the equivalent of using #[musli(packed)] on a field.

This is only implemented for type where the default behavior is not to pack the value already, this applies to types which implements EncodePacked and DecodePacked.

Examples

use musli::{Allocator, Decode, Decoder};
use musli::compat::Packed;

struct Struct {
    field: u8,
    field2: u32,
}

impl<'de, M, A> Decode<'de, M, A> for Struct
where
    A: Allocator,
    Packed<(u8, u32)>: Decode<'de, M, A>
{
    #[inline]
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de, Mode = M, Allocator = A>,
    {
        let Packed((field, field2)) = decoder.decode()?;

        Ok(Struct {
            field,
            field2,
        })
    }
}

Tuple Fields
0: T