musli::context
Struct Same

pub struct Same<M, E, A>
where
    E: ContextError,
{ /* private fields */ }

A simple non-diagnostical capturing context which simply emits the original error.

Using this should result in code which essentially just uses the emitted error type directly.

Implementations

impl<M, E> Same<M, E, &'static System>
where E: ContextError,

pub fn new() -> Self

Construct a new same-error context with a custom allocator.

impl<M, E, A> Same<M, E, A>
where E: ContextError,

pub fn with_alloc(alloc: A) -> Self

Construct a new Same context with a custom allocator.