musli::context

Trait ContextError

pub trait ContextError:
    Sized
    + 'static
    + Send
    + Sync
    + Display
    + Debug {
    // Required methods
    fn custom<T>(error: T) -> Self
       where T: 'static + Send + Sync + Error;
    fn message<T>(message: T) -> Self
       where T: Display;
}

Trait governing errors raised during encodeing or decoding.

Required Methods

fn custom<T>(error: T) -> Self
where T: 'static + Send + Sync + Error,

Construct a custom error.

fn message<T>(message: T) -> Self
where T: Display,

Collect an error from something that can be displayed.

This is made available to format custom error messages in no_std environments. The error message is to be collected by formatting T.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.