Module musli::de

Traits for generically dealing with a decoding framework.

The central traits are Decode and Decoder.

A type implementing Decode can use an Decoder to decode an instance of itself. This also comes with a derive allowing you to derive high performance decoding associated with native Rust types.

use musli::Decode;

#[derive(Decode)]
pub struct Person<'a> {
    name: &'a str,
    age: u32,
}

Enums

LengthHint

A length hint.

NumberHint

A number hint.

TypeHint

A type hint.

Traits

AsDecoder

Trait that allows a type to be repeatedly coerced into a decoder.

Decode

Trait governing how types are decoded.

DecodeOwned

Decode to a 'static value.

Decoder

Trait governing the implementation of a decoder.

NumberVisitor

A visitor capable of processing arbitrary number values.

PackDecoder

A pack that can construct encoders.

PairDecoder

Trait governing how to decode a field.

PairsDecoder

Trait governing how to decode a sequence of pairs.

SequenceDecoder

Trait governing how to decode a sequence.

ValueVisitor

A visitor for data where it might be possible to borrow it without copying from the underlying Decoder.

VariantDecoder

Trait governing how to decode a variant.

Derive Macros

Decode

Please refer to the main musli documentation.