musli::de

Trait DecodeBytes

pub trait DecodeBytes<'de, M, A>: Sized
where
    A: Allocator,
{
    const DECODE_BYTES_PACKED: bool = false;

    // Required method
    fn decode_bytes<D>(decoder: D) -> Result<Self, <D as Decoder<'de>>::Error>
       where D: Decoder<'de, Mode = M, Allocator = A>;
}

Trait governing how types are decoded as bytes.

This is typically used automatically through the #[musli(bytes)] attribute through the Decode derive.

Examples

use musli::Decode;

#[derive(Decode)]
struct MyType {
    #[musli(bytes)]
    data: [u8; 128],
}

Implementing manually:

use musli::{Allocator, Decode, Decoder};
use musli::de::DecodeBytes;

struct MyType {
    data: [u8; 128],
}

impl<'de, M, A> Decode<'de, M, A> for MyType
where
    A: Allocator,
{
    #[inline]
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: DecodeBytes::decode_bytes(decoder)?,
        })
    }
}

Provided Associated Constants

const DECODE_BYTES_PACKED: bool = false

Whether the type is packed. Packed types can be bitwise copied if the representation of the serialization format is identical to the memory layout of the type.

Note that setting this to true has safety implications, since it implies that assuming the type is correctly aligned it can be validly bitwise copied when encoded. Setting it to false is always safe.

Required Methods

fn decode_bytes<D>(decoder: D) -> Result<Self, <D as Decoder<'de>>::Error>
where D: Decoder<'de, Mode = M, Allocator = A>,

Decode the given input as bytes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.