Trait musli::de::DecodeTrace

pub trait DecodeTrace<'de, M>: Sized {
    // Required method
    fn trace_decode<D>(
        cx: &<D as Decoder<'de>>::Cx,
        decoder: D,
    ) -> Result<Self, <D as Decoder<'de>>::Error>
       where D: Decoder<'de, Mode = M>;
}

Trait governing how types are decoded specifically for tracing.

This is used for types where some extra bounds might be necessary to trace a container such as a HashMap<K, V> where K would have to implement fmt::Display.

Required Methods

fn trace_decode<D>( cx: &<D as Decoder<'de>>::Cx, decoder: D, ) -> Result<Self, <D as Decoder<'de>>::Error>
where D: Decoder<'de, Mode = M>,

Decode the given input.

Object Safety

This trait is not object safe.