musli::de
Trait DecodeUnsized

pub trait DecodeUnsized<'de, M> {
    // Required method
    fn decode_unsized<D, F, O>(
        cx: &<D as Decoder<'de>>::Cx,
        decoder: D,
        f: F,
    ) -> Result<O, <D as Decoder<'de>>::Error>
       where D: Decoder<'de, Mode = M>,
             F: FnOnce(&Self) -> Result<O, <D as Decoder<'de>>::Error>;
}

A trait implemented for types which can only be decoded by reference.

This is used for types like str which are unsized, and might require internal allocating to properly decode. Simply using the Decode implementation would restrict it to only be used through &'de str which would demand an exact reference to data from the decoded source.

Required Methods

fn decode_unsized<D, F, O>( cx: &<D as Decoder<'de>>::Cx, decoder: D, f: F, ) -> Result<O, <D as Decoder<'de>>::Error>
where D: Decoder<'de, Mode = M>, F: FnOnce(&Self) -> Result<O, <D as Decoder<'de>>::Error>,

Decode the given input using a closure as visitor.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.