Trait musli::de::Decoder

pub trait Decoder<'de>: Sized {
    type Error: Error;
    type Buffer: AsDecoder<Error = Self::Error>;
    type Some: Decoder<'de, Error = Self::Error>;
    type Pack: PackDecoder<'de, Error = Self::Error>;
    type Sequence: SequenceDecoder<'de, Error = Self::Error>;
    type Tuple: PackDecoder<'de, Error = Self::Error>;
    type Map: PairsDecoder<'de, Error = Self::Error>;
    type Struct: PairsDecoder<'de, Error = Self::Error>;
    type Variant: VariantDecoder<'de, Error = Self::Error>;

    // Required method
    fn expecting(&self, f: &mut Formatter<'_>) -> Result;

    // Provided methods
    fn type_hint(&mut self) -> Result<TypeHint, Self::Error> { ... }
    fn decode_buffer<M>(self) -> Result<Self::Buffer, Self::Error>
       where M: Mode { ... }
    fn decode_unit(self) -> Result<(), Self::Error> { ... }
    fn decode_bool(self) -> Result<bool, Self::Error> { ... }
    fn decode_char(self) -> Result<char, Self::Error> { ... }
    fn decode_u8(self) -> Result<u8, Self::Error> { ... }
    fn decode_u16(self) -> Result<u16, Self::Error> { ... }
    fn decode_u32(self) -> Result<u32, Self::Error> { ... }
    fn decode_u64(self) -> Result<u64, Self::Error> { ... }
    fn decode_u128(self) -> Result<u128, Self::Error> { ... }
    fn decode_i8(self) -> Result<i8, Self::Error> { ... }
    fn decode_i16(self) -> Result<i16, Self::Error> { ... }
    fn decode_i32(self) -> Result<i32, Self::Error> { ... }
    fn decode_i64(self) -> Result<i64, Self::Error> { ... }
    fn decode_i128(self) -> Result<i128, Self::Error> { ... }
    fn decode_usize(self) -> Result<usize, Self::Error> { ... }
    fn decode_isize(self) -> Result<isize, Self::Error> { ... }
    fn decode_f32(self) -> Result<f32, Self::Error> { ... }
    fn decode_f64(self) -> Result<f64, Self::Error> { ... }
    fn decode_number<V>(self, _: V) -> Result<V::Ok, V::Error>
       where V: NumberVisitor<Error = Self::Error> { ... }
    fn decode_array<const N: usize>(self) -> Result<[u8; N], Self::Error> { ... }
    fn decode_bytes<V>(self, _: V) -> Result<V::Ok, V::Error>
       where V: ValueVisitor<'de, Target = [u8], Error = Self::Error> { ... }
    fn decode_string<V>(self, _: V) -> Result<V::Ok, V::Error>
       where V: ValueVisitor<'de, Target = str, Error = Self::Error> { ... }
    fn decode_option(self) -> Result<Option<Self::Some>, Self::Error> { ... }
    fn decode_pack(self) -> Result<Self::Pack, Self::Error> { ... }
    fn decode_sequence(self) -> Result<Self::Sequence, Self::Error> { ... }
    fn decode_tuple(self, len: usize) -> Result<Self::Tuple, Self::Error> { ... }
    fn decode_map(self) -> Result<Self::Map, Self::Error> { ... }
    fn decode_struct(self, _: usize) -> Result<Self::Struct, Self::Error> { ... }
    fn decode_variant(self) -> Result<Self::Variant, Self::Error> { ... }

}

Trait governing the implementation of a decoder.

Required Associated Types

type Error: Error

Error type raised by the decoder.

type Buffer: AsDecoder<Error = Self::Error>

The type returned when the decoder is buffered.

type Some: Decoder<'de, Error = Self::Error>

Decoder for a value that is present.

type Pack: PackDecoder<'de, Error = Self::Error>

Pack decoder implementation.

type Sequence: SequenceDecoder<'de, Error = Self::Error>

Sequence decoder implementation.

type Tuple: PackDecoder<'de, Error = Self::Error>

Tuple decoder implementation.

type Map: PairsDecoder<'de, Error = Self::Error>

Map decoder implementation.

type Struct: PairsDecoder<'de, Error = Self::Error>

Decoder for a struct.

The caller receives a PairsDecoder which when advanced with PairsDecoder::next indicates the fields of the structure.

type Variant: VariantDecoder<'de, Error = Self::Error>

Decoder for a variant.

The caller receives a PairDecoder which when advanced with PairDecoder::first indicates which variant is being decoded and PairDecoder::second is the content of the variant.

Required Methods

fn expecting(&self, f: &mut Formatter<'_>) -> Result

Format the human-readable message that should occur if the decoder was expecting to decode some specific kind of value.

use std::fmt;

use musli::de::Decoder;
use musli::never::Never;

struct MyDecoder;

impl Decoder<'_> for MyDecoder {
    type Error = String;
    type Buffer = Never<Self::Error>;
    type Pack = Never<Self::Error>;
    type Sequence = Never<Self::Error>;
    type Tuple = Never<Self::Error>;
    type Map = Never<Self::Error>;
    type Some = Never<Self::Error>;
    type Struct = Never<Self::Error>;
    type Variant = Never<Self::Error>;

    fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        write!(f, "32-bit unsigned integers")
    }

    fn decode_u32(self) -> Result<u32, Self::Error> {
        Ok(42)
    }
}

Provided Methods

fn type_hint(&mut self) -> Result<TypeHint, Self::Error>

Return a TypeHint indicating which type is being produced by the Decoder.

Not all formats support type hints, and they might be ranging from detailed (a 32-bit unsigned integer) to vague (a number).

This is used to construct dynamic containers of types.

fn decode_buffer<M>(self) -> Result<Self::Buffer, Self::Error>
where M: Mode,

Buffer the current decoder into a buffer that can be used multiple times.

Buffering a decoder is necessary when additional introspection is needed to decode a type, but it also means that:

The entire contents of the decoder needs to be dynamically buffered in memory.

The in-memory representation might be lossy in some trivial ways. Such as arbitrary precision numbers being punted into a 64-bit float.

Examples

use musli::{Decode, Decoder, Mode};
use musli::de::{AsDecoder, PairsDecoder, PairDecoder};
use musli::error::Error;

#[derive(Decode)]
struct Variant2 {
    name: String,
    age: u32,
}

enum MyVariantType {
    Variant1,
    Variant2(Variant2),
}

impl<'de, M> Decode<'de, M> for MyVariantType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        let mut buffer = decoder.decode_buffer::<M>()?;

        let mut st = buffer.as_decoder()?.decode_map()?;

        let mut discriminator = None::<u32>;

        while let Some(mut e) = st.next()? {
            let found = e.first()?.decode_string(musli::utils::visit_string_fn(|string| {
                Ok(string == "type")
            }))?;

            if found {
                discriminator = Some(e.second().and_then(Decode::<M>::decode)?);
                break;
            }
        }

        st.end()?;

        match discriminator {
            Some(0) => Ok(MyVariantType::Variant1),
            Some(1) => Ok(MyVariantType::Variant2(buffer.as_decoder().and_then(Decode::<M>::decode)?)),
            Some(other) => Err(D::Error::invalid_variant_tag("MyVariantType", other)),
            None => Err(D::Error::missing_variant_tag("MyVariantType")),
        }
    }
}

fn decode_unit(self) -> Result<(), Self::Error>

Decode a unit or something that is empty.

Examples

use musli::{Decode, Decoder, Mode};

struct UnitType;

impl<'de, M> Decode<'de, M> for UnitType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        decoder.decode_unit()?;
        Ok(UnitType)
    }
}

fn decode_bool(self) -> Result<bool, Self::Error>

Decode a boolean.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: bool,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_bool()?,
        })
    }
}

fn decode_char(self) -> Result<char, Self::Error>

Decode a character.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: char,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_char()?,
        })
    }
}

fn decode_u8(self) -> Result<u8, Self::Error>

Decode a 8-bit unsigned integer (a.k.a. a byte).

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: u8,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_u8()?,
        })
    }
}

fn decode_u16(self) -> Result<u16, Self::Error>

Decode a 16-bit unsigned integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: u16,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_u16()?,
        })
    }
}

fn decode_u32(self) -> Result<u32, Self::Error>

Decode a 32-bit unsigned integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: u32,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_u32()?,
        })
    }
}

fn decode_u64(self) -> Result<u64, Self::Error>

Decode a 64-bit unsigned integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: u64,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_u64()?,
        })
    }
}

fn decode_u128(self) -> Result<u128, Self::Error>

Decode a 128-bit unsigned integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: u128,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_u128()?,
        })
    }
}

fn decode_i8(self) -> Result<i8, Self::Error>

Decode a 8-bit signed integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: i8,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_i8()?,
        })
    }
}

fn decode_i16(self) -> Result<i16, Self::Error>

Decode a 16-bit signed integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: i16,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_i16()?,
        })
    }
}

fn decode_i32(self) -> Result<i32, Self::Error>

Decode a 32-bit signed integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: i32,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_i32()?,
        })
    }
}

fn decode_i64(self) -> Result<i64, Self::Error>

Decode a 64-bit signed integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: i64,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_i64()?,
        })
    }
}

fn decode_i128(self) -> Result<i128, Self::Error>

Decode a 128-bit signed integer.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: i128,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_i128()?,
        })
    }
}

fn decode_usize(self) -> Result<usize, Self::Error>

Decode Rusts usize.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: usize,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_usize()?,
        })
    }
}

fn decode_isize(self) -> Result<isize, Self::Error>

Decode Rusts isize.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: isize,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_isize()?,
        })
    }
}

fn decode_f32(self) -> Result<f32, Self::Error>

Decode a 32-bit floating point value.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: f32,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_f32()?,
        })
    }
}

fn decode_f64(self) -> Result<f64, Self::Error>

Decode a 64-bit floating point value.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: f64,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_f64()?,
        })
    }
}

fn decode_number<V>(self, _: V) -> Result<V::Ok, V::Error>
where V: NumberVisitor<Error = Self::Error>,

Decode an unknown number using a visitor that can handle arbitrary precision numbers.

fn decode_array<const N: usize>(self) -> Result<[u8; N], Self::Error>

Decode a fixed-length array.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: [u8; 128],
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode_array()?,
        })
    }
}

fn decode_bytes<V>(self, _: V) -> Result<V::Ok, V::Error>
where V: ValueVisitor<'de, Target = [u8], Error = Self::Error>,

Decode a sequence of bytes whos length is encoded in the payload.

Examples

use std::fmt;
use std::marker;

use musli::de::{Decode, Decoder, ValueVisitor};
use musli::error::Error;
use musli::mode::Mode;

struct BytesReference<'de> {
    data: &'de [u8],
}

impl<'de, M> Decode<'de, M> for BytesReference<'de> where M: Mode {
    #[inline]
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        return Ok(Self {
            data: decoder.decode_bytes(Visitor(marker::PhantomData))?,
        });

        struct Visitor<E>(marker::PhantomData<E>);

        impl<'de, E> ValueVisitor<'de> for Visitor<E>
        where
            E: Error,
        {
            type Target = [u8];
            type Ok = &'de [u8];
            type Error = E;

            #[inline]
            fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
                write!(f, "exact bytes reference")
            }

            #[inline]
            fn visit_borrowed(self, bytes: &'de [u8]) -> Result<Self::Ok, Self::Error> {
                Ok(bytes)
            }
        }
    }
}

fn decode_string<V>(self, _: V) -> Result<V::Ok, V::Error>
where V: ValueVisitor<'de, Target = str, Error = Self::Error>,

Decode a string slice from the current decoder.

Examples

use std::fmt;
use std::marker;

use musli::de::{Decode, Decoder, ValueVisitor};
use musli::error::Error;
use musli::mode::Mode;

struct StringReference<'de> {
    data: &'de str,
}

impl<'de, M> Decode<'de, M> for StringReference<'de> where M: Mode {
    #[inline]
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        return Ok(Self {
            data: decoder.decode_string(Visitor(marker::PhantomData))?,
        });

        struct Visitor<E>(marker::PhantomData<E>);

        impl<'de, E> ValueVisitor<'de> for Visitor<E>
        where
            E: Error,
        {
            type Target = str;
            type Ok = &'de str;
            type Error = E;

            #[inline]
            fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
                write!(f, "exact bytes reference")
            }

            #[inline]
            fn visit_borrowed(self, bytes: &'de str) -> Result<Self::Ok, Self::Error> {
                Ok(bytes)
            }
        }
    }
}

fn decode_option(self) -> Result<Option<Self::Some>, Self::Error>

Decode an optional value.

Examples

use musli::{Decode, Decoder, Mode};

struct MyType {
    data: Option<String>,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        let data = if let Some(decoder) = decoder.decode_option()? {
            Some(<String as Decode<M>>::decode(decoder)?)
        } else {
            None
        };

        Ok(Self {
            data,
        })
    }
}

fn decode_pack(self) -> Result<Self::Pack, Self::Error>

Construct an unpack that can decode more than one element at a time.

This hints to the format that it should attempt to decode all of the elements in the packed sequence from an as compact format as possible compatible with what’s being returned by Encoder::pack.

use musli::de::{Decode, Decoder, PackDecoder};
use musli::mode::Mode;

struct PackedStruct {
    field: u32,
    data: [u8; 364],
}

impl<'de, M> Decode<'de, M> for PackedStruct where M: Mode {
    #[inline]
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        let mut unpack = decoder.decode_pack()?;
        let field = unpack.next().and_then(Decode::<M>::decode)?;
        let data = unpack.next().and_then(Decode::<M>::decode)?;
        unpack.end()?;

        Ok(Self {
            field,
            data,
        })
    }
}

fn decode_sequence(self) -> Result<Self::Sequence, Self::Error>

Decode a sequence.

Examples

use musli::de::{Decode, Decoder, SequenceDecoder};
use musli::mode::Mode;

struct MyType {
    data: Vec<String>,
}

impl<'de, M> Decode<'de, M> for MyType where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        let mut seq = decoder.decode_sequence()?;
        let mut data = Vec::new();

        while let Some(decoder) = seq.next()? {
            data.push(<String as Decode<M>>::decode(decoder)?);
        }

        seq.end()?;

        Ok(Self {
            data
        })
    }
}

fn decode_tuple(self, len: usize) -> Result<Self::Tuple, Self::Error>

Decode a fixed-length sequence of elements of length len.

Examples

use std::collections::HashMap;

use musli::de::{Decode, Decoder, PackDecoder};
use musli::error::Error;
use musli::mode::Mode;

struct TupleStruct(String, u32);

impl<'de, M> Decode<'de, M> for TupleStruct where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        let mut tuple = decoder.decode_tuple(2)?;
        let string = tuple.next().and_then(<String as Decode<M>>::decode)?;
        let integer = tuple.next().and_then(<u32 as Decode<M>>::decode)?;
        tuple.end()?;
        Ok(Self(string, integer))
    }
}

fn decode_map(self) -> Result<Self::Map, Self::Error>

Decode a map of unknown length.

The length of the map must somehow be determined from the underlying format.

Examples

use std::collections::HashMap;

use musli::de::{Decode, Decoder, PairsDecoder, PairDecoder};
use musli::mode::Mode;

struct MapStruct {
    data: HashMap<String, u32>,
}

impl<'de, M> Decode<'de, M> for MapStruct where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        let mut map = decoder.decode_map()?;
        let mut data = HashMap::with_capacity(map.size_hint().unwrap_or_default());

        while let Some(mut entry) = map.next()? {
            let key = entry.first().and_then(<String as Decode<M>>::decode)?;
            let value = entry.second().and_then(<u32 as Decode<M>>::decode)?;
            data.insert(key, value);
        }

        map.end()?;

        Ok(Self {
            data
        })
    }
}

fn decode_struct(self, _: usize) -> Result<Self::Struct, Self::Error>

Decode a struct which has an expected len number of elements.

The len indicates how many fields the decoder is expecting depending on how many fields are present in the underlying struct being decoded, butit should only be considered advisory.

The size of a struct might therefore change from one session to another.

Examples

use std::collections::HashMap;

use musli::de::{Decode, Decoder, PairsDecoder, PairDecoder};
use musli::error::Error;
use musli::mode::Mode;

struct Struct {
    string: String,
    integer: u32,
}

impl<'de, M> Decode<'de, M> for Struct where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        let mut st = decoder.decode_map()?;
        let mut string = None;
        let mut integer = None;

        while let Some(mut entry) = st.next()? {
            // Note: to avoid allocating `decode_string` needs to be used with a visitor.
            let tag = entry.first().and_then(<String as Decode<M>>::decode)?;

            match tag.as_str() {
                "string" => {
                    string = Some(entry.second().and_then(<String as Decode<M>>::decode)?);
                }
                "integer" => {
                    integer = Some(entry.second().and_then(<u32 as Decode<M>>::decode)?);
                }
                tag => {
                    return Err(D::Error::invalid_field_tag("Struct", tag))
                }
            }
        }

        st.end()?;

        Ok(Self {
            string: string.ok_or_else(|| D::Error::expected_tag("Struct", "string"))?,
            integer: integer.ok_or_else(|| D::Error::expected_tag("Struct", "integer"))?,
        })
    }
}

fn decode_variant(self) -> Result<Self::Variant, Self::Error>

Return decoder for a variant.

Examples

use musli::de::{Decode, Decoder, VariantDecoder, PairsDecoder};
use musli::error::Error;
use musli::mode::Mode;

enum Enum {
    Variant1(u32),
    Variant2(String),
}

impl<'de, M> Decode<'de, M> for Enum where M: Mode {
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        let mut variant = decoder.decode_variant()?;
        let tag = variant.tag().and_then(<usize as Decode<M>>::decode)?;

        let this = match tag {
            0 => {
                Self::Variant1(variant.variant().and_then(<u32 as Decode<M>>::decode)?)
            }
            1 => {
                Self::Variant2(variant.variant().and_then(<String as Decode<M>>::decode)?)
            }
            tag => {
                return Err(D::Error::invalid_variant_tag("Enum", tag));
            }
        };

        variant.end()?;
        Ok(this)
    }
}

Object Safety

This trait is not object safe.