musli::de

Trait EntryDecoder

pub trait EntryDecoder<'de> {
    type Cx: Context<Error = Self::Error, Allocator = Self::Allocator>;
    type Error;
    type Allocator: Allocator;
    type Mode: 'static;
    type DecodeKey<'this>: Decoder<'de, Cx = Self::Cx, Error = Self::Error, Allocator = Self::Allocator, Mode = Self::Mode>
       where Self: 'this;
    type DecodeValue: Decoder<'de, Cx = Self::Cx, Error = Self::Error, Allocator = Self::Allocator, Mode = Self::Mode>;

    // Required methods
    fn cx(&self) -> Self::Cx;
    fn decode_key(&mut self) -> Result<Self::DecodeKey<'_>, Self::Error>;
    fn decode_value(self) -> Result<Self::DecodeValue, Self::Error>;

    // Provided method
    fn size_hint(&self) -> SizeHint { ... }
}

Trait governing how to decode a map entry.

Required Associated Types

type Cx: Context<Error = Self::Error, Allocator = Self::Allocator>

Context associated with the decoder.

type Error

Error associated with decoding.

type Allocator: Allocator

The allocator associated with the decoder.

type Mode: 'static

The mode of the decoder.

type DecodeKey<'this>: Decoder<'de, Cx = Self::Cx, Error = Self::Error, Allocator = Self::Allocator, Mode = Self::Mode> where Self: 'this

The decoder to use for a tuple field index.

type DecodeValue: Decoder<'de, Cx = Self::Cx, Error = Self::Error, Allocator = Self::Allocator, Mode = Self::Mode>

The decoder to use for a tuple field value.

Required Methods

fn cx(&self) -> Self::Cx

Access the context associated with the decoder.

fn decode_key(&mut self) -> Result<Self::DecodeKey<'_>, Self::Error>

Return the decoder for the first value in the pair.

If this is a map the first value would be the key of the map, if this is a struct the first value would be the field of the struct.

fn decode_value(self) -> Result<Self::DecodeValue, Self::Error>

Decode the second value in the pair..

Provided Methods

fn size_hint(&self) -> SizeHint

Get a size hint for the size of the map being decoded.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.