Trait musli::de::NumberVisitor

pub trait NumberVisitor: Sized {
    type Ok;
    type Error: Error;

    // Required method
    fn expecting(&self, f: &mut Formatter<'_>) -> Result;

    // Provided methods
    fn visit_u8(self, _: u8) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_u16(self, _: u16) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_u32(self, _: u32) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_u64(self, _: u64) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_u128(self, _: u128) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_i8(self, _: i8) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_i16(self, _: i16) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_i32(self, _: i32) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_i64(self, _: i64) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_i128(self, _: i128) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_f32(self, _: f32) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_f64(self, _: f64) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_usize(self, _: usize) -> Result<Self::Ok, Self::Error> { ... }
    fn visit_isize(self, _: isize) -> Result<Self::Ok, Self::Error> { ... }

}

A visitor capable of processing arbitrary number values.

Required Associated Types

type Ok

The output of the visitor.

type Error: Error

An error type.

Required Methods

fn expecting(&self, f: &mut Formatter<'_>) -> Result

Format an error indicating what was expected by this visitor.

Override to be more specific about the type that failed.

Provided Methods

fn visit_u8(self, _: u8) -> Result<Self::Ok, Self::Error>

Visit u8.

fn visit_u16(self, _: u16) -> Result<Self::Ok, Self::Error>

Visit u16.

fn visit_u32(self, _: u32) -> Result<Self::Ok, Self::Error>

Visit u32.

fn visit_u64(self, _: u64) -> Result<Self::Ok, Self::Error>

Visit u64.

fn visit_u128(self, _: u128) -> Result<Self::Ok, Self::Error>

Visit u128.

fn visit_i8(self, _: i8) -> Result<Self::Ok, Self::Error>

Visit i8.

fn visit_i16(self, _: i16) -> Result<Self::Ok, Self::Error>

Visit i16.

fn visit_i32(self, _: i32) -> Result<Self::Ok, Self::Error>

Visit i32.

fn visit_i64(self, _: i64) -> Result<Self::Ok, Self::Error>

Visit i64.

fn visit_i128(self, _: i128) -> Result<Self::Ok, Self::Error>

Visit i128.

fn visit_f32(self, _: f32) -> Result<Self::Ok, Self::Error>

Visit f32.

fn visit_f64(self, _: f64) -> Result<Self::Ok, Self::Error>

Visit f64.

fn visit_usize(self, _: usize) -> Result<Self::Ok, Self::Error>

Visit usize.

fn visit_isize(self, _: isize) -> Result<Self::Ok, Self::Error>

Visit isize.