Trait musli::de::SequenceDecoder

pub trait SequenceDecoder<'de>: Sized {
    type Cx: Context + ?Sized;
    type DecodeNext<'this>: Decoder<'de, Cx = Self::Cx, Error = <Self::Cx as Context>::Error, Mode = <Self::Cx as Context>::Mode>
       where Self: 'this;

    // Required methods
    fn decode_next(
        &mut self
    ) -> Result<Self::DecodeNext<'_>, <Self::Cx as Context>::Error>;
    fn try_decode_next(
        &mut self
    ) -> Result<Option<Self::DecodeNext<'_>>, <Self::Cx as Context>::Error>;

    // Provided methods
    fn size_hint(&self) -> SizeHint { ... }
    fn next<T>(&mut self) -> Result<T, <Self::Cx as Context>::Error>
       where Self: Sized,
             T: Decode<'de, <Self::Cx as Context>::Mode> { ... }
    fn try_next<T>(&mut self) -> Result<Option<T>, <Self::Cx as Context>::Error>
       where Self: Sized,
             T: Decode<'de, <Self::Cx as Context>::Mode> { ... }
}

Trait governing how to decode a sequence.

Required Associated Types

type Cx: Context + ?Sized

Context associated with the decoder.

type DecodeNext<'this>: Decoder<'de, Cx = Self::Cx, Error = <Self::Cx as Context>::Error, Mode = <Self::Cx as Context>::Mode> where Self: 'this

The decoder for individual items.

Required Methods

fn decode_next( &mut self ) -> Result<Self::DecodeNext<'_>, <Self::Cx as Context>::Error>

Return decoder to decode the next element.

This will error or provide garbled data in case the next element is not available.

fn try_decode_next( &mut self ) -> Result<Option<Self::DecodeNext<'_>>, <Self::Cx as Context>::Error>

Try to decode the next element.

Provided Methods

fn size_hint(&self) -> SizeHint

Get a size hint of known remaining elements.

fn next<T>(&mut self) -> Result<T, <Self::Cx as Context>::Error>
where Self: Sized, T: Decode<'de, <Self::Cx as Context>::Mode>,

Decode the next element of the given type, erroring in case it’s absent.

fn try_next<T>(&mut self) -> Result<Option<T>, <Self::Cx as Context>::Error>
where Self: Sized, T: Decode<'de, <Self::Cx as Context>::Mode>,

Decode the next element of the given type.

Object Safety

This trait is not object safe.