Trait musli::de::VariantDecoder

pub trait VariantDecoder<'de> {
    type Error: Error;
    type Tag<'this>: Decoder<'de, Error = Self::Error>
       where Self: 'this;
    type Variant<'this>: Decoder<'de, Error = Self::Error>
       where Self: 'this;

    // Required methods
    fn tag(&mut self) -> Result<Self::Tag<'_>, Self::Error>;
    fn variant(&mut self) -> Result<Self::Variant<'_>, Self::Error>;
    fn skip_variant(&mut self) -> Result<bool, Self::Error>;
    fn end(self) -> Result<(), Self::Error>;
}

Trait governing how to decode a variant.

Required Associated Types

type Error: Error

Error type.

type Tag<'this>: Decoder<'de, Error = Self::Error> where Self: 'this

The decoder to use for the variant tag.

type Variant<'this>: Decoder<'de, Error = Self::Error> where Self: 'this

The decoder to use for the variant value.

Required Methods

fn tag(&mut self) -> Result<Self::Tag<'_>, Self::Error>

Return the decoder for the first value in the pair.

If this is a map the first value would be the key of the map, if this is a struct the first value would be the field of the struct.

fn variant(&mut self) -> Result<Self::Variant<'_>, Self::Error>

Decode the second value in the pair..

fn skip_variant(&mut self) -> Result<bool, Self::Error>

Indicate that the second value should be skipped.

The boolean returned indicates if the value was skipped or not.

fn end(self) -> Result<(), Self::Error>

End the pair decoder.

Object Safety

This trait is not object safe.