Trait musli::en::EncodePacked

pub trait EncodePacked<M> {
    // Required method
    fn encode_packed<E>(
        &self,
        cx: &<E as Encoder>::Cx,
        encoder: E,
    ) -> Result<<E as Encoder>::Ok, <E as Encoder>::Error>
       where E: Encoder<Mode = M>;
}

Trait governing how a type is encoded as a packed value.

Packed encodings are ones where data follow one after another, with no “metadata” indicating when one value starts and another stops.

This is typically used automatically through the #[musli(packed)] attribute through the Decode derive.

Examples

use musli::Encode;

#[derive(Encode)]
struct PackedType {
    #[musli(packed)]
    data: (u32, u32),
}

Implementing manually:

use musli::{Encode, Encoder};
use musli::en::SequenceEncoder;

struct PackedType {
    data: (u32, u32),
}

impl<M> Encode<M> for PackedType {
    fn encode<E>(&self, cx: &E::Cx, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder,
    {
        let mut pack = encoder.encode_pack()?;
        pack.push(&self.data.0);
        pack.push(&self.data.1);
        pack.finish_sequence()
    }
}

Required Methods

fn encode_packed<E>( &self, cx: &<E as Encoder>::Cx, encoder: E, ) -> Result<<E as Encoder>::Ok, <E as Encoder>::Error>
where E: Encoder<Mode = M>,

Encode the given output as bytes.

Object Safety

This trait is not object safe.