Trait musli::en::Encoder

pub trait Encoder: Sized {
    type Ok;
    type Error: Error;
    type Some: Encoder<Ok = Self::Ok, Error = Self::Error>;
    type Pack: SequenceEncoder<Ok = Self::Ok, Error = Self::Error>;
    type Sequence: SequenceEncoder<Ok = Self::Ok, Error = Self::Error>;
    type Tuple: SequenceEncoder<Ok = Self::Ok, Error = Self::Error>;
    type Map: PairsEncoder<Ok = Self::Ok, Error = Self::Error>;
    type Struct: PairsEncoder<Ok = Self::Ok, Error = Self::Error>;
    type Variant: VariantEncoder<Ok = Self::Ok, Error = Self::Error>;

    // Required method
    fn expecting(&self, f: &mut Formatter<'_>) -> Result;

    // Provided methods
    fn encode_unit(self) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_bool(self, _: bool) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_char(self, _: char) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_u8(self, _: u8) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_u16(self, _: u16) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_u32(self, _: u32) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_u64(self, _: u64) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_u128(self, _: u128) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_i8(self, _: i8) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_i16(self, _: i16) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_i32(self, _: i32) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_i64(self, _: i64) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_i128(self, _: i128) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_usize(self, _: usize) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_isize(self, _: isize) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_f32(self, _: f32) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_f64(self, _: f64) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_array<const N: usize>(
        self,
        _: [u8; N]
    ) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_bytes(self, _: &[u8]) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_bytes_vectored(self, _: &[&[u8]]) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_string(self, _: &str) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_some(self) -> Result<Self::Some, Self::Error> { ... }
    fn encode_none(self) -> Result<Self::Ok, Self::Error> { ... }
    fn encode_pack(self) -> Result<Self::Pack, Self::Error> { ... }
    fn encode_sequence(self, len: usize) -> Result<Self::Sequence, Self::Error> { ... }
    fn encode_tuple(self, len: usize) -> Result<Self::Tuple, Self::Error> { ... }
    fn encode_map(self, len: usize) -> Result<Self::Map, Self::Error> { ... }
    fn encode_struct(self, _: usize) -> Result<Self::Struct, Self::Error> { ... }
    fn encode_variant(self) -> Result<Self::Variant, Self::Error> { ... }

}

Trait governing how the encoder works.

Required Associated Types

type Ok

The type returned by the encoder. For Encode implementations ensures that they are used correctly, since only functions returned by the Encoder is capable of returning this value.

type Error: Error

The error raised by an encoder.

type Some: Encoder<Ok = Self::Ok, Error = Self::Error>

Encoder returned when encoding an optional value which is present.

type Pack: SequenceEncoder<Ok = Self::Ok, Error = Self::Error>

A simple pack that packs a sequence of elements.

type Sequence: SequenceEncoder<Ok = Self::Ok, Error = Self::Error>

The type of a sequence encoder.

type Tuple: SequenceEncoder<Ok = Self::Ok, Error = Self::Error>

The type of a tuple encoder.

type Map: PairsEncoder<Ok = Self::Ok, Error = Self::Error>

The type of a map encoder.

type Struct: PairsEncoder<Ok = Self::Ok, Error = Self::Error>

Encoder that can encode a struct.

type Variant: VariantEncoder<Ok = Self::Ok, Error = Self::Error>

Encoder for a struct variant.

Required Methods

fn expecting(&self, f: &mut Formatter<'_>) -> Result

An expectation error. Every other implementation defers to this to report that something unexpected happened.

Provided Methods

fn encode_unit(self) -> Result<Self::Ok, Self::Error>

Encode a unit or something that is completely empty.

Examples

use musli::{Encode, Encoder, Mode};

struct EmptyStruct;

impl<M> Encode<M> for EmptyStruct where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_unit()
    }
}

fn encode_bool(self, _: bool) -> Result<Self::Ok, Self::Error>

Encode a boolean value.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: bool,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_bool(self.data)
    }
}

fn encode_char(self, _: char) -> Result<Self::Ok, Self::Error>

Encode a character.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: char,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_char(self.data)
    }
}

fn encode_u8(self, _: u8) -> Result<Self::Ok, Self::Error>

Encode a 8-bit unsigned integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: u8,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_u8(self.data)
    }
}

fn encode_u16(self, _: u16) -> Result<Self::Ok, Self::Error>

Encode a 16-bit unsigned integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: u16,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_u16(self.data)
    }
}

fn encode_u32(self, _: u32) -> Result<Self::Ok, Self::Error>

Encode a 32-bit unsigned integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: u32,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_u32(self.data)
    }
}

fn encode_u64(self, _: u64) -> Result<Self::Ok, Self::Error>

Encode a 64-bit unsigned integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: u64,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_u64(self.data)
    }
}

fn encode_u128(self, _: u128) -> Result<Self::Ok, Self::Error>

Encode a 128-bit unsigned integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: u128,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_u128(self.data)
    }
}

fn encode_i8(self, _: i8) -> Result<Self::Ok, Self::Error>

Encode a 8-bit signed integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: i8,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_i8(self.data)
    }
}

fn encode_i16(self, _: i16) -> Result<Self::Ok, Self::Error>

Encode a 16-bit signed integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: i16,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_i16(self.data)
    }
}

fn encode_i32(self, _: i32) -> Result<Self::Ok, Self::Error>

Encode a 32-bit signed integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: i32,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_i32(self.data)
    }
}

fn encode_i64(self, _: i64) -> Result<Self::Ok, Self::Error>

Encode a 64-bit signed integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: i64,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_i64(self.data)
    }
}

fn encode_i128(self, _: i128) -> Result<Self::Ok, Self::Error>

Encode a 128-bit signed integer.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: i128,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_i128(self.data)
    }
}

fn encode_usize(self, _: usize) -> Result<Self::Ok, Self::Error>

Encode Rusts usize.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: usize,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_usize(self.data)
    }
}

fn encode_isize(self, _: isize) -> Result<Self::Ok, Self::Error>

Encode Rusts isize.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: isize,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_isize(self.data)
    }
}

fn encode_f32(self, _: f32) -> Result<Self::Ok, Self::Error>

Encode a 32-bit floating point value.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: f32,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_f32(self.data)
    }
}

fn encode_f64(self, _: f64) -> Result<Self::Ok, Self::Error>

Encode a 64-bit floating point value.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: f64,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_f64(self.data)
    }
}

fn encode_array<const N: usize>( self, _: [u8; N] ) -> Result<Self::Ok, Self::Error>

Encode fixed-length array.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: [u8; 364],
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_array(self.data)
    }
}

fn encode_bytes(self, _: &[u8]) -> Result<Self::Ok, Self::Error>

Encode a sequence of bytes.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: Vec<u8>,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_bytes(self.data.as_slice())
    }
}

fn encode_bytes_vectored(self, _: &[&[u8]]) -> Result<Self::Ok, Self::Error>

Encode the given slices of bytes in sequence, with one following another as a single contiguous byte array.

This can be useful to avoid allocations when a caller doesn’t have access to a single byte sequence like in VecDeque.

Examples

use std::collections::VecDeque;

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: VecDeque<u8>,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        let (first, second) = self.data.as_slices();
        encoder.encode_bytes_vectored(&[first, second])
    }
}

fn encode_string(self, _: &str) -> Result<Self::Ok, Self::Error>

Encode a string.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: String,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        encoder.encode_string(self.data.as_str())
    }
}

fn encode_some(self) -> Result<Self::Some, Self::Error>

Encode an optional value that is present.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: Option<String>,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        match &self.data {
            Some(data) => {
                encoder.encode_some().and_then(|e| Encode::<M>::encode(data, e))
            }
            None => {
                encoder.encode_none()
            }
        }
    }
}

fn encode_none(self) -> Result<Self::Ok, Self::Error>

Encode an optional value that is absent.

Examples

use musli::{Encode, Encoder, Mode};

struct MyType {
    data: Option<String>,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        match &self.data {
            Some(data) => {
                encoder.encode_some().and_then(|e| Encode::<M>::encode(data, e))
            }
            None => {
                encoder.encode_none()
            }
        }
    }
}

fn encode_pack(self) -> Result<Self::Pack, Self::Error>

Construct a pack that can encode more than one element at a time.

This hints to the format that it should attempt to encode all of the elements in the packed sequence as compact as possible and that subsequent unpackers will know the exact length of the element being unpacked.

Examples

use musli::en::{Encode, Encoder, SequenceEncoder};
use musli::mode::Mode;

struct PackedStruct {
    field: u32,
    data: [u8; 364],
}

impl<M> Encode<M> for PackedStruct where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        let mut pack = encoder.encode_pack()?;
        pack.next()?.encode_u32(self.field)?;
        pack.next()?.encode_array(self.data)?;
        pack.end()
    }
}

fn encode_sequence(self, len: usize) -> Result<Self::Sequence, Self::Error>

Encode a sequence with a known length len.

A sequence encodes one element following another and must in some way encode the length of the sequence in the underlying format. It is decoded with Decoder::decode_sequence.

Examples

use musli::en::{Encode, Encoder, SequenceEncoder};
use musli::mode::Mode;

struct MyType {
    data: Vec<String>,
}

impl<M> Encode<M> for MyType where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        let mut seq = encoder.encode_sequence(self.data.len())?;

        for element in &self.data {
            seq.push::<M, _>(element)?;
        }

        seq.end()
    }
}

fn encode_tuple(self, len: usize) -> Result<Self::Tuple, Self::Error>

Encode a tuple with a known length len.

This is almost identical to Encoder::encode_sequence except that we know that we are encoding a fixed-length container of length len, and assuming the size of that container doesn’t change in size it can be decoded using Decoder::decode_tuple again without the underlying format having to encode the size of the container.

Examples

use musli::en::{Encode, Encoder, SequenceEncoder};
use musli::mode::Mode;

struct PackedTuple(u32, [u8; 364]);

impl<M> Encode<M> for PackedTuple where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        let mut tuple = encoder.encode_tuple(2)?;
        tuple.next()?.encode_u32(self.0)?;
        tuple.next()?.encode_array(self.1)?;
        tuple.end()
    }
}

fn encode_map(self, len: usize) -> Result<Self::Map, Self::Error>

Encode a map with a known length len.

fn encode_struct(self, _: usize) -> Result<Self::Struct, Self::Error>

Encode a struct.

Examples

use musli::en::{Encode, Encoder, PairEncoder, PairsEncoder};
use musli::mode::Mode;

struct Struct {
    name: String,
    age: u32,
}

impl<M> Encode<M> for Struct where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        let mut st = encoder.encode_struct(2)?;
        st.insert::<M, _, _>("name", &self.name)?;
        st.insert::<M, _, _>("age", self.age)?;
        st.end()
    }
}

fn encode_variant(self) -> Result<Self::Variant, Self::Error>

Encode an struct enum variant.

Examples

use musli::en::{Encode, Encoder, VariantEncoder, PairsEncoder};
use musli::mode::Mode;

enum Enum {
    UnitVariant,
    TupleVariant(String),
    Variant {
        data: String,
        age: u32,
    }
}

impl<M> Encode<M> for Enum where M: Mode {
    fn encode<E>(&self, encoder: E) -> Result<E::Ok, E::Error>
    where
        E: Encoder
    {
        let mut variant = encoder.encode_variant()?;

        match self {
            Enum::UnitVariant => {
                variant.insert::<M, _, _>("variant1", ())
            }
            Enum::TupleVariant(data) => {
                variant.insert::<M, _, _>("variant2", data)
            }
            Enum::Variant { data, age } => {
                variant.tag()?.encode_string("variant3")?;

                let mut st = variant.variant()?.encode_struct(2)?;
                st.insert::<M, _, _>("data", data)?;
                st.insert::<M, _, _>("age", age)?;
                st.end()?;

                variant.end()
            }
        }
    }
}

Object Safety

This trait is not object safe.