Trait musli::error::Error

pub trait Error: Sized + 'static + Send + Sync + Display + Debug {
    // Required methods
    fn custom<T>(error: T) -> Self
       where T: 'static + Send + Sync + Display + Debug;
    fn message<T>(message: T) -> Self
       where T: Display;

    // Provided methods
    fn uninhabitable(type_name: &'static str) -> Self { ... }
    fn expected_tag<T>(type_name: &'static str, tag: T) -> Self
       where T: Debug { ... }
    fn invalid_field_tag<T>(type_name: &'static str, tag: T) -> Self
       where T: Debug { ... }
    fn invalid_variant_tag<T>(type_name: &'static str, tag: T) -> Self
       where T: Debug { ... }
    fn missing_variant_field<T>(type_name: &'static str, tag: T) -> Self
       where T: Debug { ... }
    fn missing_variant_tag(type_name: &'static str) -> Self { ... }
    fn invalid_variant_field_tag<V, T>(
        type_name: &'static str,
        variant: V,
        tag: T
    ) -> Self
       where V: Debug,
             T: Debug { ... }
}

Trait governing errors raised during encodeing or decoding.

Required Methods

fn custom<T>(error: T) -> Selfwhere T: 'static + Send + Sync + Display + Debug,

Construct a custom error.

fn message<T>(message: T) -> Selfwhere T: Display,

Collect an error from something that can be displayed.

This is made available to format custom error messages in no_std environments. The error message is to be collected by formatting T.

Provided Methods

fn uninhabitable(type_name: &'static str) -> Self

Trying to decode an uninhabitable type.

fn expected_tag<T>(type_name: &'static str, tag: T) -> Selfwhere T: Debug,

The value for the given tag could not be collected.

fn invalid_field_tag<T>(type_name: &'static str, tag: T) -> Selfwhere T: Debug,

Encountered an unsupported number tag.

fn invalid_variant_tag<T>(type_name: &'static str, tag: T) -> Selfwhere T: Debug,

Indicate that a variant wasn’t supported by tag.

fn missing_variant_field<T>(type_name: &'static str, tag: T) -> Selfwhere T: Debug,

Missing variant field required to decode.

fn missing_variant_tag(type_name: &'static str) -> Self

Indicate that a variant tag could not be determined.

fn invalid_variant_field_tag<V, T>( type_name: &'static str, variant: V, tag: T ) -> Selfwhere V: Debug, T: Debug,

Encountered an unsupported variant field.