musli::hint

Trait SequenceHint

pub trait SequenceHint: Sized + Sealed {
    // Required method
    fn get(self) -> Option<usize>;

    // Provided methods
    fn require<C>(self, cx: C) -> Result<usize, <C as Context>::Error>
       where C: Context { ... }
    fn size_hint(self) -> SizeHint { ... }
}

A size hint passed in when encoding or decoding a sequence.

Required Methods

fn get(self) -> Option<usize>

Get an optional sequence hint.

Provided Methods

fn require<C>(self, cx: C) -> Result<usize, <C as Context>::Error>
where C: Context,

Require a sequence hint or raise an error indicating that the format doesn’t support sequences of unknown size if one is not present.

fn size_hint(self) -> SizeHint

Coerce into a [SizeHint].

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.