musli::reader
Trait Reader

pub trait Reader<'de>: Sealed {
    type Mut<'this>: Reader<'de>
       where Self: 'this;

    // Required methods
    fn borrow_mut(&mut self) -> Self::Mut<'_>;
    fn skip<C>(&mut self, cx: &C, n: usize) -> Result<(), C::Error>
       where C: ?Sized + Context;
    fn peek(&mut self) -> Option<u8>;
    fn read_bytes<C, V>(
        &mut self,
        cx: &C,
        n: usize,
        visitor: V,
    ) -> Result<V::Ok, C::Error>
       where C: ?Sized + Context,
             V: UnsizedVisitor<'de, C, [u8]>;

    // Provided methods
    fn read<C>(&mut self, cx: &C, buf: &mut [u8]) -> Result<(), C::Error>
       where C: ?Sized + Context { ... }
    fn read_byte<C>(&mut self, cx: &C) -> Result<u8, C::Error>
       where C: ?Sized + Context { ... }
    fn read_array<C, const N: usize>(
        &mut self,
        cx: &C,
    ) -> Result<[u8; N], C::Error>
       where C: ?Sized + Context { ... }
    fn limit(self, limit: usize) -> Limit<Self>
       where Self: Sized { ... }
}

Trait governing how a source of bytes is read.

This requires the reader to be able to hand out contiguous references to the byte source through Reader::read_bytes.

Required Associated Types

type Mut<'this>: Reader<'de> where Self: 'this

Type borrowed from self.

Why oh why would we want to do this over having a simple &'this mut T?

We want to avoid recursive types, which will blow up the compiler. And the above is a typical example of when that can go wrong. This ensures that each call to borrow_mut dereferences the Reader at each step to avoid constructing a large muted type, like &mut &mut &mut SliceReader<'de>.

Required Methods

fn borrow_mut(&mut self) -> Self::Mut<'_>

Borrow the current reader.

fn skip<C>(&mut self, cx: &C, n: usize) -> Result<(), C::Error>
where C: ?Sized + Context,

Skip over the given number of bytes.

fn peek(&mut self) -> Option<u8>

Peek the next value.

fn read_bytes<C, V>( &mut self, cx: &C, n: usize, visitor: V, ) -> Result<V::Ok, C::Error>
where C: ?Sized + Context, V: UnsizedVisitor<'de, C, [u8]>,

Read a slice out of the current reader.

Provided Methods

fn read<C>(&mut self, cx: &C, buf: &mut [u8]) -> Result<(), C::Error>
where C: ?Sized + Context,

Read a slice into the given buffer.

fn read_byte<C>(&mut self, cx: &C) -> Result<u8, C::Error>
where C: ?Sized + Context,

Read a single byte.

fn read_array<C, const N: usize>(&mut self, cx: &C) -> Result<[u8; N], C::Error>
where C: ?Sized + Context,

Read an array out of the current reader.

fn limit(self, limit: usize) -> Limit<Self>
where Self: Sized,

Keep an accurate record of the position within the reader.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.