Trait musli::Allocator

pub trait Allocator {
    type Buf<'this>: Buf
       where Self: 'this;

    // Required method
    fn alloc(&self) -> Option<Self::Buf<'_>>;
}

An allocator that can be used in combination with a context.

See the musli-allocator crate for examples.

Required Associated Types

type Buf<'this>: Buf where Self: 'this

The type of an allocated buffer.

Required Methods

fn alloc(&self) -> Option<Self::Buf<'_>>

Allocate an empty, uninitialized buffer.

Calling this method returns None if the allocation failed.

Object Safety

This trait is not object safe.