musli

Trait Decode

pub trait Decode<'de, M, A>: Sized
where
    A: Allocator,
{
    const IS_BITWISE_DECODE: bool = false;

    // Required method
    fn decode<D>(decoder: D) -> Result<Self, <D as Decoder<'de>>::Error>
       where D: Decoder<'de, Mode = M, Allocator = A>;
}

Trait governing how types are decoded.

This is typically implemented automatically using the Decode derive.

Examples

use musli::Decode;

#[derive(Decode)]
struct MyType {
    data: [u32; 8],
}

Implementing manually:

use musli::{Allocator, Decode, Decoder};

struct MyType {
    data: [u32; 8],
}

impl<'de, M, A> Decode<'de, M, A> for MyType
where
    A: Allocator,
{
    #[inline]
    fn decode<D>(decoder: D) -> Result<Self, D::Error>
    where
        D: Decoder<'de>,
    {
        Ok(Self {
            data: decoder.decode()?,
        })
    }
}

Provided Associated Constants

const IS_BITWISE_DECODE: bool = false

Whether the type is packed. Packed types can be bitwise copied if the representation of the serialization format is identical to the memory layout of the type.

Note that setting this to true has safety implications, since it implies that assuming the type is correctly aligned it can be validly bitwise copied when encoded. Setting it to false is always safe.

This being set to true also implies that the type is Copy, and must not have a Drop implementation.

Required Methods

fn decode<D>(decoder: D) -> Result<Self, <D as Decoder<'de>>::Error>
where D: Decoder<'de, Mode = M, Allocator = A>,

Decode the current value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.