Module musli::compat

Wrapper types for tweaking how something is encoded.

Note that most types in this module have an attribute equivalent:

Bytes corresponds to using #[musli(bytes)] on a field.

Packed corresponds to using #[musli(packed)] on a field.

Structs

Bytes

Treat T as if its bytes.

Packed

Treat T as if its packed.

Sequence

Ensures that the given value T is encoded as a sequence.