musli
Module context

Context implementations.

Structs

Capture

A simple non-diagnostical capturing context.

DefaultContext

The default context which uses an allocator to track the location of errors.

Error

A collected error which has been context decorated.

ErrorMarker

Indicates that an error occurred during encoding. This is a placeholder error that can be used by context implementations and is a ZST.

Ignore

A simple non-diagnostical capturing context which ignores the error and loses all information about it (except that it happened).

Same

A simple non-diagnostical capturing context which simply emits the original error.

Traits

ContextError

Trait governing errors raised during encodeing or decoding.

Functions

new

Construct a new default context using the static System allocator.

with_alloc

Construct a new default context using the provided allocator.