Module musli::context

Context implementations.

Structs

Capture

A simple non-diagnostical capturing context.

ErrorMarker

Indicates that an error occurred during encoding. This is a placeholder error that can be used by context implementations and is a ZST.

Ignore

A simple non-diagnostical capturing context which ignores the error and loses all information about it (except that it happened).

RichError

A collected error which has been context decorated.

Same

A simple non-diagnostical capturing context which simply emits the original error.

StackContext

A rich context which uses allocations and tracks the exact location of errors.

SystemContext

A rich context dynamically allocating space using the system allocator.

Traits

Error

Trait governing errors raised during encodeing or decoding.