musli

Module context

Context implementations.

Structs

Capture

Capture an error of the specified type E.

DefaultContext

The default context which uses an allocator to track the location of errors.

Emit

Emit an error of the specified type E.

Error

A collected error which has been context decorated.

ErrorMarker

Indicates that an error occurred during encoding. This is a placeholder error that can be used by context implementations and is a ZST.

Errors

An iterator over available errors.

Ignore

Disable error capture.

NoTrace

Trace configuration indicating that tracing is fully disabled.

Report

A line-separated report of all errors.

Trace

Marker type indicating that tracing is enabled.

Traits

ContextError

Trait governing errors raised during encodeing or decoding.

ErrorMode

The trait governing how error capture is implemented.

TraceImpl

The trait governing how tracing works in a DefaultContext.

TraceMode

Trait for marker types indicating the tracing mode to use.

Functions

new

Construct a new default context using the System allocator.

new_in

Construct a new default context using the provided allocator.