musli::context

Struct DefaultContext

pub struct DefaultContext<A, T, C>
where
    A: Allocator,
    T: TraceMode,
{ /* private fields */ }

The default context which uses an allocator to track the location of errors.

This is typically constructed using new and by default uses the System allocator to allocate memory. To customized the allocator to use new_in can be used during construction.

The default constructor is only available when the alloc feature is enabled, and will use the System allocator.

Implementations

impl<A, T, C> DefaultContext<A, T, C>
where A: Allocator, T: TraceMode, C: ErrorMode<A>,

pub fn with_trace(self) -> DefaultContext<A, Trace, C>

Enable tracing through the current allocator A.

Note that this makes diagnostics methods such as report and errors available on the type.

Tracing requires the configured allocator to work, if for example the Disabled allocator was in use, no diagnostics would be collected.

pub fn with_capture<E>(self) -> DefaultContext<A, T, Capture<E>>
where E: ContextError<A>,

Capture the specified error type.

This gives access to the last captured error through DefaultContext::unwrap and DefaultContext::result.

Capturing instead of forwarding the error might be beneficial if the error type is large.

Examples

use musli::{Decode, Encode};
use musli::alloc::System;
use musli::context;
use musli::json::{Encoding, Error};

const ENCODING: Encoding = Encoding::new();

#[derive(Decode, Encode)]
struct Person {
    name: String,
    age: u32,
}

let cx = context::new().with_capture::<Error>();

let mut data = Vec::new();

ENCODING.encode_with(&cx, &mut data, &Person {
    name: "Aristotle".to_string(),
    age: 61,
})?;

assert!(cx.result().is_ok());

let _: Result<Person, _> = ENCODING.from_slice_with(&cx, &data[..data.len() - 2]);
assert!(cx.result().is_err());
Ok::<_, musli::context::ErrorMarker>(())

pub fn with_error<E>(self) -> DefaultContext<A, T, Emit<E>>
where E: ContextError<A>,

Emit the specified error type E.

This causes the method receiving the context to return the specified error type directly instead through Context::Error.

Examples

use musli::{Decode, Encode};
use musli::alloc::System;
use musli::context;
use musli::json::{Encoding, Error};

const ENCODING: Encoding = Encoding::new();

#[derive(Decode, Encode)]
struct Person {
    name: String,
    age: u32,
}

let cx = context::new().with_error();

let mut data = Vec::new();

ENCODING.encode_with(&cx, &mut data, &Person {
    name: "Aristotle".to_string(),
    age: 61,
})?;

let person: Person = ENCODING.from_slice_with(&cx, &data[..])?;
assert_eq!(person.name, "Aristotle");
assert_eq!(person.age, 61);
Ok::<_, Error>(())

impl<A, C> DefaultContext<A, Trace, C>
where A: Allocator,

pub fn with_type(self) -> Self

If tracing is enabled through DefaultContext::with_trace, this configured the context to visualize type information, and not just variant and fields.

Examples

use musli::context;

let cx = context::new().with_trace().with_type();

pub fn report(&self) -> Report<'_, A>

Generate a line-separated report of all reported errors.

This can be useful if you want a quick human-readable overview of errors. The line separator used will be platform dependent.

Examples

use musli::context::{self, ErrorMarker};
use musli::value::Value;
use musli::json::Encoding;

const ENCODING: Encoding = Encoding::new();

let cx = context::new().with_trace();

let ErrorMarker = ENCODING.from_str_with::<_, Value<_>>(&cx, "not json").unwrap_err();
let report = cx.report();
println!("{report}");

pub fn errors(&self) -> Errors<'_, A>

Iterate over all reported errors.

Examples

use musli::context::{self, ErrorMarker};
use musli::value::Value;
use musli::json::Encoding;

const ENCODING: Encoding = Encoding::new();

let cx = context::new().with_trace();

let ErrorMarker = ENCODING.from_str_with::<_, Value<_>>(&cx, "not json").unwrap_err();
assert!(cx.errors().count() > 0);

impl<A, T, E> DefaultContext<A, T, Capture<E>>
where A: Allocator, T: TraceMode, E: ContextError<A>,

pub fn unwrap(&self) -> E

Unwrap the error marker or panic if there is no error.

pub fn result(&self) -> Result<(), E>

Coerce a captured error into a result.