Module musli::de

Traits for generically dealing with a decoding framework.

The central traits are Decode and Decoder.

A type implementing Decode can use an Decoder to decode an instance of itself. This also comes with a derive allowing you to derive high performance decoding associated with native Rust types.

use musli::Decode;

#[derive(Decode)]
pub struct Person<'a> {
    name: &'a str,
    age: u32,
}

Structs

SizeHint

A length hint.

Enums

Skip

Indicates if skipping was performed.

Traits

AsDecoder

Trait that allows a type to be repeatedly coerced into a decoder.

Decode

Trait governing how types are decoded.

DecodeBytes

Trait governing how types are decoded as bytes.

DecodeOwned

Decode to an owned value.

DecodePacked

Trait governing how a type is decoded as a packed value.

DecodeTrace

Trait governing how types are decoded specifically for tracing.

DecodeUnsized

A trait implemented for types which can only be decoded by reference.

DecodeUnsizedBytes

A trait implemented for types which can be visited by reference.

Decoder

Trait governing the implementation of a decoder.

EntriesDecoder

Trait governing how to decode a sequence of map pairs.

EntryDecoder

Trait governing how to decode a map entry.

MapDecoder

Trait governing how to decode a sequence of pairs.

SequenceDecoder

Trait governing how to decode a sequence.

UnsizedVisitor

A visitor for data where we might need to borrow without copying from the underlying Decoder.

VariantDecoder

Trait governing how to decode a variant.

Visitor

Visitor capable of decoding any type into a value Visitor::Ok.

Derive Macros

Decode

Derive which automatically implements the Decode trait.