Trait musli::de::PairDecoder

pub trait PairDecoder<'de> {
    type Error: Error;
    type First<'this>: Decoder<'de, Error = Self::Error>
       where Self: 'this;
    type Second: Decoder<'de, Error = Self::Error>;

    // Required methods
    fn first(&mut self) -> Result<Self::First<'_>, Self::Error>;
    fn second(self) -> Result<Self::Second, Self::Error>;
    fn skip_second(self) -> Result<bool, Self::Error>;
}

Trait governing how to decode a field.

Required Associated Types

type Error: Error

Error type.

type First<'this>: Decoder<'de, Error = Self::Error> where Self: 'this

The decoder to use for a tuple field index.

type Second: Decoder<'de, Error = Self::Error>

The decoder to use for a tuple field value.

Required Methods

fn first(&mut self) -> Result<Self::First<'_>, Self::Error>

Return the decoder for the first value in the pair.

If this is a map the first value would be the key of the map, if this is a struct the first value would be the field of the struct.

fn second(self) -> Result<Self::Second, Self::Error>

Decode the second value in the pair..

fn skip_second(self) -> Result<bool, Self::Error>

Indicate that the second value should be skipped.

The boolean returned indicates if the value was skipped or not.

Object Safety

This trait is not object safe.