Trait musli::en::VariantEncoder

pub trait VariantEncoder {
    type Ok;
    type Error: Error;
    type Tag<'this>: Encoder<Ok = Self::Ok, Error = Self::Error>
       where Self: 'this;
    type Variant<'this>: Encoder<Ok = Self::Ok, Error = Self::Error>
       where Self: 'this;

    // Required methods
    fn tag(&mut self) -> Result<Self::Tag<'_>, Self::Error>;
    fn variant(&mut self) -> Result<Self::Variant<'_>, Self::Error>;
    fn end(self) -> Result<Self::Ok, Self::Error>;

    // Provided method
    fn insert<M, F, S>(
        self,
        first: F,
        second: S
    ) -> Result<Self::Ok, Self::Error>
       where Self: Sized,
             M: Mode,
             F: Encode<M>,
             S: Encode<M> { ... }
}

Trait governing how to encode a variant.

Required Associated Types

type Ok

Result type of the encoder.

type Error: Error

The error raised by a map encoder.

type Tag<'this>: Encoder<Ok = Self::Ok, Error = Self::Error> where Self: 'this

The encoder returned when advancing the map encoder to encode the key.

type Variant<'this>: Encoder<Ok = Self::Ok, Error = Self::Error> where Self: 'this

The encoder returned when advancing the map encoder to encode the value.

Required Methods

fn tag(&mut self) -> Result<Self::Tag<'_>, Self::Error>

Return the encoder for the first element in the variant.

fn variant(&mut self) -> Result<Self::Variant<'_>, Self::Error>

Return encoder for the second element in the variant.

fn end(self) -> Result<Self::Ok, Self::Error>

End the variant encoder.

Provided Methods

fn insert<M, F, S>(self, first: F, second: S) -> Result<Self::Ok, Self::Error>
where Self: Sized, M: Mode, F: Encode<M>, S: Encode<M>,

Insert the variant immediately.

Object Safety

This trait is not object safe.