musli

Function is_bitwise_encode

pub const fn is_bitwise_encode<T>() -> bool
where
    T: Encode<Binary>,

Test if the given type T is marked with the #[musli(packed)] attribute, and is bitwise encodeable in the Binary mode.

See the help section for #[musli(packed)] for more information.