musli::reader
Struct Limit

pub struct Limit<R> { /* private fields */ }

Limit the number of bytes that can be read out of a reader to the specified limit.

Constructed through Reader::limit.

Implementations

impl<R> Limit<R>

pub fn remaining(&self) -> usize

Get the remaining data in the limited reader.