musli

Module storage

Efficient binary storage encoding for Müsli.

The storage encoding is partially upgrade safe:

✔ Can tolerate missing fields if they are annotated with #[musli(default)].

✗ Cannot skip over extra unrecognized fields.

This means that it’s suitable as a storage format, since the data model only evolves in one place. But unsuitable as a wire format since it cannot allow clients to upgrade independent of each other.

See wire or descriptive for formats which are upgrade stable.

use musli::{Encode, Decode};

#[derive(Debug, PartialEq, Encode, Decode)]
struct Version1 {
    name: String,
}

#[derive(Debug, PartialEq, Encode, Decode)]
struct Version2 {
    name: String,
    #[musli(default)]
    age: Option<u32>,
}

let version2 = musli::storage::to_vec(&Version2 {
    name: String::from("Aristotle"),
    age: Some(61),
})?;

assert!(musli::storage::decode::<_, Version1>(version2.as_slice()).is_err());

let version1 = musli::storage::to_vec(&Version1 {
    name: String::from("Aristotle"),
})?;

let version2: Version2 = musli::storage::decode(version1.as_slice())?;

assert_eq!(version2, Version2 {
    name: String::from("Aristotle"),
    age: None,
});

Configuring

To tweak the behavior of the storage format you can use the Encoding type:

use musli::{Encode, Decode};
use musli::mode::Binary;
use musli::options::{self, Options, Integer};
use musli::storage::Encoding;

const OPTIONS: Options = options::new().integer(Integer::Fixed).build();
const CONFIG: Encoding<OPTIONS> = Encoding::new().with_options();

#[derive(Debug, PartialEq, Encode, Decode)]
struct Person<'a> {
    name: &'a str,
    age: u32,
}

let mut out = Vec::new();

let expected = Person {
    name: "Aristotle",
    age: 61,
};

CONFIG.encode(&mut out, &expected)?;
let actual = CONFIG.decode(&out[..])?;

assert_eq!(expected, actual);

Structs

Encoding

Setting up encoding with parameters.

Error

Error raised during storage encoding.

Constants

OPTIONS

The default options for the storage encoding.

Functions

decode

Decode the given type T from the given Reader using the default Encoding.

encode

Encode the given value to the given Writer using the default Encoding.

from_slice

Decode the given type T from the given slice using the default Encoding.

to_fixed_bytes

Encode the given value to a fixed-size bytes using the default Encoding.

to_slice

Encode the given value to the given slice using the default Encoding and return the number of bytes encoded.

to_vec

Encode the given value to a Vec using the default Encoding.

to_writer

Encode the given value to the given Write using the default Encoding.

Type Aliases

Result

Convenient result alias for use with musli::storage.