musli::writer
Trait Writer

pub trait Writer {
    type Mut<'this>: Writer
       where Self: 'this;

    // Required methods
    fn borrow_mut(&mut self) -> Self::Mut<'_>;
    fn extend<C>(
        &mut self,
        cx: &C,
        buffer: Vec<'_, u8, C::Allocator>,
    ) -> Result<(), C::Error>
       where C: ?Sized + Context;
    fn write_bytes<C>(&mut self, cx: &C, bytes: &[u8]) -> Result<(), C::Error>
       where C: ?Sized + Context;

    // Provided method
    fn write_byte<C>(&mut self, cx: &C, b: u8) -> Result<(), C::Error>
       where C: ?Sized + Context { ... }
}

The trait governing how a writer works.

Required Associated Types

type Mut<'this>: Writer where Self: 'this

Reborrowed type.

Why oh why would we want to do this over having a simple &'this mut T?

We want to avoid recursive types, which will blow up the compiler. And the above is a typical example of when that can go wrong. This ensures that each call to borrow_mut dereferences the Reader at each step to avoid constructing a large muted type, like &mut &mut &mut VecWriter.

Required Methods

fn borrow_mut(&mut self) -> Self::Mut<'_>

Reborrow the current type.

fn extend<C>( &mut self, cx: &C, buffer: Vec<'_, u8, C::Allocator>, ) -> Result<(), C::Error>
where C: ?Sized + Context,

Write a buffer to the current writer.

fn write_bytes<C>(&mut self, cx: &C, bytes: &[u8]) -> Result<(), C::Error>
where C: ?Sized + Context,

Write bytes to the current writer.

Provided Methods

fn write_byte<C>(&mut self, cx: &C, b: u8) -> Result<(), C::Error>
where C: ?Sized + Context,

Write a single byte.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.